Move the ManifestLoader class into the manifest.
This allows it to be shared by code outside wptrunner that wants to
be compatible with loading multiple manifests.
