+++ tools/wptrunner/wptrunner/testloader.py
@@ -12,18 +12,38 @@ import manifestexpected
 import wpttest
 from mozlog import structured
 
-manifest = None
-manifest_update = None
-download_from_github = None
-manifest_log = None
+manifest_loader = None
+
 
 def do_delayed_imports():
-    # This relies on an already loaded module having set the sys.path correctly :(
-    global manifest, manifest_update, download_from_github, manifest_log
-    from manifest import manifest
-    from manifest import update as manifest_update
-    from manifest.download import download_from_github
-    from manifest import log as manifest_log
+    global manifest_loader
+    from manifest import loader as manifest_loader
+
+
+# This function is here for backwards compat with the old class
+def ManifestLoader(test_paths, force_manifest_update=False, manifest_download=False,
+                   types=None, meta_filters=None, logger=None):
+    do_delayed_imports()
+    if logger is None:
+        logger = structured.get_default_logger()
+    if logger is None:
+        logger = structured.structuredlog.StructuredLogger("wptrunner")
+    return manifest_loader.ManifestLoader(logger,
+                                          test_paths,
+                                          force_manifest_update=force_manifest_update,
+                                          manifest_download=manifest_download,
+                                          types=types,
+                                          meta_filters=meta_filters)
+
+
+def load_manifests(test_paths, force_manifest_update=False, manifest_download=False,
+                   types=None, meta_filters=None, logger=None):
+    return ManifestLoader(test_paths,
+                          force_manifest_update=force_manifest_update,
+                          manifest_download=manifest_download,
+                          types=types,
+                          meta_filters=meta_filters,
+                          logger=logger).load()
 
 
 class TestChunker(object):
@@ -367,6 +387,7 @@ class TestFilter(object):
             if include_tests:
                 yield test_type, test_path, include_tests
 
+
 class TagFilter(object):
     def __init__(self, tags):
         self.tags = set(tags)
@@ -375,78 +396,6 @@ class TagFilter(object):
         for test in test_iter:
             if test.tags & self.tags:
                 yield test
-
-
-class ManifestLoader(object):
-    def __init__(self, test_paths, force_manifest_update=False, manifest_download=False, types=None, meta_filters=None):
-        do_delayed_imports()
-        self.test_paths = test_paths
-        self.force_manifest_update = force_manifest_update
-        self.manifest_download = manifest_download
-        self.types = types
-        self.meta_filters = meta_filters or []
-        self.logger = structured.get_default_logger()
-        if self.logger is None:
-            self.logger = structured.structuredlog.StructuredLogger("ManifestLoader")
-
-    def load(self):
-        rv = {}
-        for url_base, paths in self.test_paths.iteritems():
-            manifest_file = self.load_manifest(url_base=url_base,
-                                               **paths)
-            path_data = {"url_base": url_base}
-            path_data.update(paths)
-            rv[manifest_file] = path_data
-        return rv
-
-    def create_manifest(self, manifest_path, tests_path, url_base="/"):
-        self.update_manifest(manifest_path, tests_path, url_base, recreate=True,
-                             download=self.manifest_download)
-
-    def update_manifest(self, manifest_path, tests_path, url_base="/",
-                        recreate=False, download=False):
-        self.logger.info("Updating test manifest %s" % manifest_path)
-        manifest_log.setup()
-
-        json_data = None
-        if download:
-            # TODO: make this not github-specific
-            download_from_github(manifest_path, tests_path)
-
-        if not recreate:
-            try:
-                with open(manifest_path) as f:
-                    json_data = json.load(f)
-            except IOError:
-                self.logger.info("Unable to find test manifest")
-            except ValueError:
-                self.logger.info("Unable to parse test manifest")
-
-        if not json_data:
-            self.logger.info("Creating test manifest")
-            manifest_file = manifest.Manifest(url_base)
-        else:
-            try:
-                manifest_file = manifest.Manifest.from_json(tests_path, json_data)
-            except manifest.ManifestVersionMismatch:
-                manifest_file = manifest.Manifest(url_base)
-
-        manifest_update.update(tests_path, manifest_file, True)
-
-        manifest.write(manifest_file, manifest_path)
-
-    def load_manifest(self, tests_path, manifest_path, url_base="/", **kwargs):
-        if (not os.path.exists(manifest_path) or
-            self.force_manifest_update):
-            self.update_manifest(manifest_path, tests_path, url_base, download=self.manifest_download)
-        manifest_file = manifest.load(tests_path, manifest_path, types=self.types, meta_filters=self.meta_filters)
-        if manifest_file.url_base != url_base:
-            self.logger.info("Updating url_base in manifest from %s to %s" % (manifest_file.url_base,
-                                                                              url_base))
-            manifest_file.url_base = url_base
-            manifest.write(manifest_file, manifest_path)
-
-        return manifest_file
 
 
 def iterfilter(filters, iter):
